backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
